refactor: 레포지토리 계층 추상화
