+++ src/main/java/roomescape/repository/ReservationTimeRepositoryImpl.java
@@ -1,0 +1,58 @@
+package roomescape.repository;
+
+import java.util.List;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.dto.CreateReservationTimeDto;
+import roomescape.entity.ReservationTime;
+
+@Repository
+public class ReservationTimeRepositoryImpl implements ReservationTimeRepository {
+    private final JdbcTemplate jdbcTemplate;
+
+    public ReservationTimeRepositoryImpl(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Long addAndGetId(CreateReservationTimeDto createReservationTimeDto) {
+        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate.getDataSource())
+                .withTableName("reservation_time")
+                .usingColumns("start_at")
+                .usingGeneratedKeyColumns("id");
+
+        SqlParameterSource parameters = new MapSqlParameterSource()
+                .addValue("start_at", createReservationTimeDto.startAt());
+
+        return simpleJdbcInsert.executeAndReturnKey(parameters).longValue();
+    }
+
+    @Override
+    public ReservationTime findById(Long id) {
+        String sql = "SELECT id, start_at FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql,
+                (resultSet, rowNum) -> new ReservationTime(
+                        resultSet.getLong("id"),
+                        resultSet.getString("start_at")
+                ), id);
+    }
+
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT id, start_at FROM reservation_time";
+        return jdbcTemplate.query(sql,
+                (resultSet, rowNum) -> new ReservationTime(
+                        resultSet.getLong("id"),
+                        resultSet.getString("start_at")
+                ));
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+        jdbcTemplate.update(sql, id);
+    }
+}
